Move stream and schema into separate namespace
