+++ aeon/schema/foraging.py
@@ -1,6 +1,7 @@
 import pandas as _pd
 import aeon.io.reader as _reader
 import aeon.io.device as _device
+import aeon.schema.core as _stream
 from enum import Enum as _Enum
 
 class Area(_Enum):
@@ -21,14 +22,6 @@ class _RegionReader(_reader.Harp):
         data['region'] = categorical.rename_categories(Area._member_names_)
         return data
 
-def video(pattern):
-    """Video frame metadata."""
-    return { "Video": _reader.Video(pattern) }
-
-def position(pattern):
-    """Position tracking data for the specified camera."""
-    return { "Position": _reader.Position(f"{pattern}_200") }
-
 def region(pattern):
     """Region tracking data for the specified camera."""
     return { "Region": _RegionReader(f"{pattern}_201") }
@@ -36,10 +29,6 @@ def region(pattern):
 def depletionFunction(pattern):
     """State of the linear depletion function for foraging patches."""
     return { "DepletionState": _reader.PatchState(f"{pattern}_State") }
-
-def encoder(pattern):
-    """Wheel magnetic encoder data."""
-    return { "Encoder": _reader.Encoder(f"{pattern}_90") }
 
 def feeder(pattern):
     """Feeder commands and events."""
@@ -55,7 +44,7 @@ def deliver_pellet(pattern):
 
 def patch(pattern):
     """Data streams for a patch."""
-    return _device.compositeStream(pattern, depletionFunction, encoder, feeder)
+    return _device.compositeStream(pattern, depletionFunction, _stream.encoder, feeder)
 
 def weight(pattern):
     """Weight measurement data streams for a specific nest."""
@@ -72,26 +61,6 @@ def weight_filtered(pattern):
 def weight_subject(pattern):
     """Subject weight measurement for a specific nest."""
     return { "WeightSubject": _reader.Weight(f"{pattern}_204") }
-
-def environment(pattern):
-    """Metadata for environment mode and subjects."""
-    return _device.compositeStream(pattern, environment_state, subject_state)
-
-def environment_state(pattern):
-    """Environment state log."""
-    return { "EnvironmentState": _reader.Csv(f"{pattern}_EnvironmentState", ['state']) }
-
-def subject_state(pattern):
-    """Subject state log."""
-    return { "SubjectState": _reader.Subject(f"{pattern}_SubjectState") }
-
-def messageLog(pattern):
-    """Message log data."""
-    return { "MessageLog": _reader.Log(f"{pattern}_MessageLog") }
-
-def metadata(pattern):
-    """Metadata for acquisition epochs."""
-    return { pattern: _reader.Metadata(pattern) }
 
 def session(pattern):
     """Session metadata for Experiment 0.1."""
